blog/static_file/requirement.txt
adium-theme-ubuntu==0.3.4
decorator==4.0.6
Django==1.11.3
flake8==2.5.4
image==1.5.9
ipython==2.4.1
mccabe==0.2.1
mysqlclient==1.3.7
olefile==0.44
pep8==1.7.0
pexpect==4.0.1
Pillow==4.2.1
ptyprocess==0.5
pyflakes==1.1.0
pytz==2017.2
simplegeneric==0.8.1
unity-lens-photos==1.0
unrar==0.3
